Changing cast place enhances speed

+++ scripts/depth_map.py
@@ -152,7 +152,7 @@ def classify_near_far(gt, bins, bin_length, err):
 	res = [[] for i in range(len(bins))]
 
 	# find bin: [0-X] -> 0, (X-2X] -> 1, ....
-	gt2 = np.floor(np.array(gt) / bin_length)
+	gt2 = np.floor(np.array(gt) / bin_length).astype(np.int)
 
 	err = np.array(err)
 
@@ -163,7 +163,7 @@ def classify_near_far(gt, bins, bin_length, err):
 
 	# apply data to its corresponding bin
 	for i in range(len(err)):
-		res[int(gt2[i])].append(err[i])
+		res[gt2[i]].append(err[i])
 
 	return np.array(map(lambda x : np.round(x, 6), res))
